requirements.txt
asyncio==3.4.3
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
